Modified special case in create_time to return system uptime for PID 0 on Windows and Linux.
I encountered two problems on Windows:

I've not been able to use time() function of <time.h> because it seems there's a strange bug in mingw that causes an "ImportError: specified procedure cannot be found" exception raised when psutil is imported:
http://bugs.python.org/issue3308

I solved this one by passing time.time() value from Python to the C function.

The other problem is that by using GetTickCount() time will wrap around to zero if the system is run continuously for 49.7 days.
I added a XXX note in the C source code to remind us that there's space for improvement.